requirements.txt
fastapi
uvicorn[standard]
httpx
supabase
python-dotenv
jinja2
aiofiles
python-multipart
